Safari-611.1.10/LayoutTests/fast/dom/constructor-proto.html

<!DOCTYPE html>
<script src="../../resources/js-test-pre.js"></script>
<script>

description('This tests that setting the proto of a built in constructor is not reset');

// Hotmail depends on the [[Prototype]] of DOM constructors to be Object.prototype.
shouldBe('Object.getPrototypeOf(HTMLElement)', 'Element');

var proto = new Number(42);
HTMLElement.__proto__ = proto;
shouldBe('Object.getPrototypeOf(HTMLElement)', 'proto');

</script>
<script src="../../resources/js-test-post.js"></script>
