third_party/blink/web_tests/platform/generic/fast/dom/constructor-proto-expected.txt

This tests that setting the proto of a built in constructor is not reset

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Object.getPrototypeOf(HTMLElement) is Element
PASS Object.getPrototypeOf(HTMLElement) is proto
PASS successfullyParsed is true

TEST COMPLETE
